events_network.c: explicit conversion to int
socketlist only increases from growsocketlist(), which is a static function
which is only called by events_network_register() after checking that the
socket number is less than FD_SETSIZE.

As a result, we can be be certain that socketlist_getsize(S) will return a
value much less than INT_MAX, so it can be safely cast to an int.

+++ libcperciva/events/events_network.c
@@ -255,7 +255,7 @@ events_network_select(struct timeval * tv)
 	events_network_selectstats_select();
 
 	/* Select. */
-	while (select(socketlist_getsize(S), &readfds, &writefds,
+	while (select((int)socketlist_getsize(S), &readfds, &writefds,
 	    NULL, tv) == -1) {
 		/* EINTR is harmless. */
 		if (errno == EINTR)
